Refactored Product entity, added interface contract and applied catalog promotion aware trait. Fixed some mapping with relations between promo group and promoted products. Added test app entities as example for overriding plugin content.
