+++ src/Entity/ProductTrait.php
@@ -1,16 +1,25 @@
 <?php
 
-declare(strict_types=1);
-
 namespace Locastic\SyliusCatalogPromotionPlugin\Entity;
 
 use Sylius\Component\Core\Model\ChannelInterface;
-use Sylius\Component\Core\Model\Product as BaseProduct;
+use Sylius\Component\Core\Model\ChannelPricingInterface;
 use Sylius\Component\Core\Model\ProductVariantInterface;
 
-class Product extends BaseProduct implements ProductInterface
+trait ProductTrait
 {
-    use CatalogPromotionGroupAwareTrait;
+    /** @var CatalogPromotionGroupInterface */
+    private $appliedCatalogPromotionGroup;
+
+    public function getAppliedCatalogPromotionGroup(): ?CatalogPromotionGroupInterface
+    {
+        return $this->appliedCatalogPromotionGroup;
+    }
+
+    public function setAppliedCatalogPromotionGroup(?CatalogPromotionGroupInterface $catalogPromotionGroup): void
+    {
+        $this->appliedCatalogPromotionGroup = $catalogPromotionGroup;
+    }
 
     public function getAppliedCatalogPromotion(): ?CatalogPromotionInterface
     {
@@ -19,15 +28,11 @@ class Product extends BaseProduct implements ProductInterface
 
     public function getPromotionDiscounts(ChannelInterface $channel)
     {
-        $promotionDiscounts = [];
-
-        $promotionDiscounts = $this->getVariants()->map(function (ProductVariantInterface $productVariant) use ($channel, $promotionDiscounts) {
+        return $this->getVariants()->map(function (ProductVariantInterface $productVariant) use ($channel) {
             /** @var ChannelPricingInterface $channelPricing */
             $channelPricing = $productVariant->getChannelPricingForChannel($channel);
 
             return $channelPricing->getDiscount();
         });
-
-        return $promotionDiscounts;
     }
 }
