--- src/Resources/config/doctrine/Product.orm.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-    <mapped-superclass name="Locastic\SyliusCatalogPromotionPlugin\Entity\Product" table="sylius_product">
-        <many-to-one field="appliedCatalogPromotionGroup" target-entity="Locastic\SyliusCatalogPromotionPlugin\Entity\CatalogPromotionGroup" inversed-by="products">
-            <join-column name="catalog_promotion_group_id" referenced-column-name="id" nullable="true" on-delete="SET NULL" />
-        </many-to-one>
-    </mapped-superclass>
-</doctrine-mapping>
