+++ src/Resources/config/resources.yml
@@ -30,11 +30,6 @@ sylius_resource:
                 model: Locastic\SyliusCatalogPromotionPlugin\Entity\CatalogBannerImage
                 form: Locastic\SyliusCatalogPromotionPlugin\Form\Type\CatalogBannerImageType
 
-sylius_product:
-    resources:
-        product:
-            classes:
-                model: Locastic\SyliusCatalogPromotionPlugin\Entity\Product
 sylius_core:
     resources:
         channel_pricing:
