+++ tests/Application/app/config/resources.yml
@@ -1,5 +1,17 @@
+sylius_resource:
+    resources:
+        locastic_sylius_catalog_promotion_plugin.catalog_promotion_group:
+            classes:
+                model: Tests\Acme\SyliusExamplePlugin\Application\src\Entity\CatalogPromotionGroup
+
 sylius_core:
     resources:
         channel_pricing:
             classes:
                 model: Tests\Acme\SyliusExamplePlugin\Application\src\Entity\ChannelPricing
+
+sylius_product:
+    resources:
+        product:
+            classes:
+                model: Tests\Acme\SyliusExamplePlugin\Application\src\Entity\Product
