+++ tests/Application/src/Entity/CatalogPromotionGroup.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Acme\SyliusExamplePlugin\Application\src\Entity;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Locastic\SyliusCatalogPromotionPlugin\Entity\CatalogPromotion;
+
+class CatalogPromotionGroup extends CatalogPromotion
+{
+    /** @var Product[]|ArrayCollection */
+    private $products;
+}
